+++ source/extensions/omni.flux.utils.material_converter/omni/flux/utils/material_converter/impl/usd_preview_surface_to_aperture_pbr.py
@@ -26,7 +26,7 @@ from omni.flux.utils.material_converter.base.converter_builder_base import Conve
 from pxr import Sdf, UsdShade
 
 if TYPE_CHECKING:
-    from pxr import Gf, Usd
+    from pxr import Usd
 
 
 class _NormalMapEncodings(Enum):
@@ -47,7 +47,6 @@ class USDPreviewSurfaceToAperturePBRConverterBuilder(ConverterBuilderBase):
                 input_attr_name="inputs:diffuseColor",
                 output_attr_name="inputs:diffuse_texture",
                 translate_fn=self._convert_connection_to_texture,
-                translate_alt_fn=self._convert_connection_to_texture_alt,
             ),
             AttributeBase(input_attr_name="inputs:opacity", output_attr_name="inputs:opacity_constant"),
             AttributeBase(
@@ -58,20 +57,17 @@ class USDPreviewSurfaceToAperturePBRConverterBuilder(ConverterBuilderBase):
                 input_attr_name="inputs:roughness",
                 output_attr_name="inputs:reflectionroughness_texture",
                 translate_fn=self._convert_connection_to_texture,
-                translate_alt_fn=self._convert_connection_to_texture_alt,
             ),
             AttributeBase(input_attr_name="inputs:metallic", output_attr_name="inputs:metallic_constant"),
             AttributeBase(
                 input_attr_name="inputs:metallic",
                 output_attr_name="inputs:metallic_texture",
                 translate_fn=self._convert_connection_to_texture,
-                translate_alt_fn=self._convert_connection_to_texture_alt,
             ),
             AttributeBase(
                 input_attr_name="inputs:displacement",
                 output_attr_name="inputs:height_texture",
                 translate_fn=self._convert_connection_to_texture,
-                translate_alt_fn=self._convert_connection_to_texture_alt,
             ),
             # Need conversion
             AttributeBase(input_attr_name="inputs:emissiveColor", output_attr_name="inputs:emissive_color_constant"),
@@ -79,7 +75,6 @@ class USDPreviewSurfaceToAperturePBRConverterBuilder(ConverterBuilderBase):
                 input_attr_name="inputs:emissiveColor",
                 output_attr_name="inputs:emissive_mask_texture",
                 translate_fn=self._convert_connection_to_texture,
-                translate_alt_fn=self._convert_connection_to_texture_alt,
             ),
             AttributeBase(input_attr_name="inputs:encoding", output_attr_name="inputs:encoding", fake_attribute=True),
         ]
@@ -89,7 +84,9 @@ class USDPreviewSurfaceToAperturePBRConverterBuilder(ConverterBuilderBase):
             attributes=attributes,
         )
 
-    def _convert_connection_to_texture(self, _value: Gf.Vec3f, input_attr: Usd.Attribute) -> Sdf.AssetPath:
+    def _convert_connection_to_texture(
+        self, _value: object, input_attr: Usd.Attribute
+    ) -> tuple[Sdf.ValueTypeName, Sdf.AssetPath]:
         """
         Convert the attribute. If the attribute has a connection, and this connection is `UsdUVTexture` with a `file`
         input, we set the texture.
@@ -99,29 +96,21 @@ class USDPreviewSurfaceToAperturePBRConverterBuilder(ConverterBuilderBase):
             input_attr: the input attr that need to be translated
 
         Returns:
-            The value to set to the attribute
+            The output attribute type and value.
         """
         connections = input_attr.GetConnections()
         shd_input = UsdShade.Input(input_attr)
         if len(connections) > 0:
             if len(connections) > 1:
-                return Sdf.AssetPath()
+                return Sdf.ValueTypeNames.Asset, Sdf.AssetPath()
             connected_source = shd_input.GetConnectedSource()
             if connected_source is None:
-                return Sdf.AssetPath()
+                return Sdf.ValueTypeNames.Asset, Sdf.AssetPath()
             # The source must be a valid shader or material prim.
             connected_shader = UsdShade.Shader(connected_source[0].GetPrim())
             if connected_shader.GetShaderId() == "UsdUVTexture":
                 file_input = connected_shader.GetInput("file")
                 outputs = connected_source[0].GetOutputs()
                 if outputs[0].GetBaseName() == "rgb" and file_input:
-                    return file_input.Get()
-        return Sdf.AssetPath()
-
-    def _convert_connection_to_texture_alt(
-        self, _: Sdf.ValueTypeNames, value: Gf.Vec3f, input_attr: Usd.Attribute | None
-    ) -> tuple[Sdf.ValueTypeNames, str]:
-        return (
-            Sdf.ValueTypeNames.Asset,
-            self._convert_connection_to_texture(value, input_attr),
-        )
+                    return Sdf.ValueTypeNames.Asset, file_input.Get()
+        return Sdf.ValueTypeNames.Asset, Sdf.AssetPath()