+++ source/extensions/omni.flux.utils.material_converter/omni/flux/utils/material_converter/tests/unit/base/test_attribute_base.py
@@ -19,31 +19,38 @@ from unittest.mock import Mock
 
 import omni.kit.test
 from omni.flux.utils.material_converter.base.attribute_base import AttributeBase
+from pxr import Sdf
 
 
 class TestAttributeBase(omni.kit.test.AsyncTestCase):
-    async def test_translate_fn_should_have_default_behavior_should_return_same_value(self):
+    async def test_translate_fn_with_default_behavior_should_return_input_type_and_value(self):
         # Arrange
-        attr = AttributeBase.construct()
-        val_mock = Mock()
-        attr_mock = Mock()
-
-        # Act
-        val = attr.translate_fn(val_mock, attr_mock)
-
-        # Assert
-        self.assertEqual(val_mock, val)
-
-    async def test_translate_alt_fn_should_have_default_behavior_should_return_same_values(self):
-        # Arrange
-        attr = AttributeBase.construct()
+        attr = AttributeBase(input_attr_name="input", output_attr_name="output")
         type_mock = Mock()
         val_mock = Mock()
         attr_mock = Mock()
+        attr_mock.GetTypeName.return_value = type_mock
 
         # Act
-        translated_type, translated_val = attr.translate_alt_fn(type_mock, val_mock, attr_mock)
+        translated_type, translated_val = attr.translate_fn(val_mock, attr_mock)
 
         # Assert
         self.assertEqual(type_mock, translated_type)
         self.assertEqual(val_mock, translated_val)
+
+    async def test_output_default_type_and_value_should_store_both(self):
+        # Arrange
+        type_mock = Sdf.ValueTypeNames.Int
+        val_mock = Mock()
+
+        # Act
+        attr = AttributeBase(
+            input_attr_name="input",
+            output_attr_name="output",
+            output_attr_type=type_mock,
+            output_default_value=val_mock,
+        )
+
+        # Assert
+        self.assertEqual(type_mock, attr.output_attr_type)
+        self.assertEqual(val_mock, attr.output_default_value)